Condensed code
Functionality that has been successfully implemented in code condense:
Spawn patron with a specific request.
Patron will properly go to stool or line.
Patron will leave if not attended to in time
Patron will leave if all resources are in use
Basic interaction with patron.

Things that were not able to be implemented just yet:
UI
Patrons traverse the line
Interact with resources.
Day Transition
Time keeping in day.