Add Game Clock
Added a script that adds a clock that counts down from 10 minutes; needs
a little tweaking, but works. Also added the clock and opponent's score
to the HUD. Removed an old script that contained static variables to
store the player and opponent's score, since Chris added a script that
replaced it. Also added the "missing" scripts to their respective game
objects.