Added Scene Fading Transitions
-Removed the arbitrary End Level scene, and added two new scenes; one to
be used for the score recap, and one for the store.
-Moved the GameClock and ScoreTracker scripts to New Scripts folder, and
deleted the HUD Scripts folder.
-Added fade out to GameClock script. Also added a Start function that
sets the clock based on the current
scene. Also will switch to the appropriate scene based on current scene.
-Changed Minute and Hour in Inn script both to 0.
-Created a script called SceneFade to fade out of a scene.
-GameClock and SceneFade added to the end of day recap scene. SceneFade
added to store scene.
-Added the two new scenes to BuildSettings.

NOTE: The scene transition from the store scene back to the main scene
is dependent on the buttons that are included in the store. So I have
not coded that yet.