Added functionality and resource state debugger.
Added a prefab for tables that includes the table and stools and can
simply be dragged and dropped

Added obstacle script to the table itself to prevent patrons from
walking through the table to get to their seat.

Added expanding resources in resource manager. Upon loading the main
scene the ResourceManager script will look up the current upgrade
values and set those game objects for each resource to true.