Tavern Layout Update
Updates the tavern layout. Adjusts the colider on the walls to look more
realistic when the player moves into the top all, adds beds and stools,
enlarges the bar in lobby and adjusts transform values on all resources
to be more uniform and pleasing. Also reenables the invisible wall to
prevent player from moving out-of-bounds from the bottom.