Build edits
-Moved the PolyNav2DInspector script into a folder named Editor in order
to fix issues preventing build.
-Reduced player movement speed (it was really fast in the build version)
-Fixed the score display's position to prevent clipping
-Moved the images in the Store scene (deleted bag of gold)
-Added the main menu scene to the build settings
-Added a zip folder containing builds for windows

BUILD NOTES: The screen in the unity editor did not accurately represent
the screen size of the build version. Therefore, the HUD had to be moved
to compensate for this. The doorways for each room are also cut off as a
result (I didn't adjust the wall assets, as I was afraid of tampering
with the pathfinding). Also, the player movement was extremely fast in
the build version and had to be reduced. Some scene transitions also
take longer than others.